Broke down get waf package dict into helper functions
